Take the map out of the page stack.
This allows returning to the last viewed menu page because pages don't
need to be popped out of stack to return to the map. Rather the page
stack is just hidden using ApplicationWindow.bottomMargin, and can thus
be trivially restored. This approach also allows a split screen, which
might come in handy one day.