Fix tile display with Sailfish OS 1.1.9 "Eineheminlampi".
Jolla seems to have updated the version of QtLocation in Sailfish to
contain a post 5.2.1 commit that changes the tile size used by the Nokia
map plugin from 256 to 512, without having any plugin parameter to
control that. This broke our tile loading hack. Fix the problem by
introducing two new constants canvasTileSize and canvasScaleFactor,
which handle the necessary transformations. These can be reset to the
standard 256 and 1 if at some later version of QtLocation we can finally
switch to different map plugin.

https://github.com/qtproject/qtlocation/commit/1ca75bd6bc109c6e5ad47fd5c236709d3f64c466