Try to avoid initial centering on new route being wrong
I can't reproduce the problem, but sometimes (very rarely) the
centering, which should center on the route and zoom so that whole route
is visible, would go in some weird corner of the planet. I'm guessing
this has to do with either the centering animation or auto-center and
auto-rotate updates while running fitViewToRoute.