Remove navigationReroutable and navigationTarget
These are not saved across sessions, so they are useless if the route
was calculated earlier. The target can be assumed to be the same as the
last point of the existing route, thus also eliminating any need to
geocode the same point again. Starting navigation to the current
position is not really relevant enough to guard against.

Add simpler rerouting property to guard against concurrent reroutes.