Add support for @3x tiles
Amend the system of zoom level locking that previously locked to integer
zoom levels to now conditionally also lock to half zoom levels,
log2(1.5) to be exact, so that @3x, @6x etc. tiles can be displayed
pixel for pixel. Zoom levels are still in various places, such as the
config file, rounded down to integer values, but since the
transformations are reversible, everything should probably work as
intended.

Closes #53