better support for convergence by route calculation, fixes #179
The route calculation page, which used to be just busy indicator,
shows maneuvers after the route has been found. This is for the
platforms supporting convergence (Kirigami). For other platforms,
this part is the same.

For all platforms, narrative page remembers its status better by
keeping a flag app.narrativePageSeen. Only if maneuvers are refreshed
(tracked by resetting the flag in Map.updateManeuvers), we populate the list.
Otherwise, narrative page stays as it was, keeping scroll position and
active maneuver