allowing to select offline maps even without osmscout installed
Without OSM Scout Server, if user selects offline profile, (s)he can get stuck into the case where none of the maps
are available for selected profile. To avoid this case and handle absence of OSM Scout Server via GUI, the corresponding
styles are allowed. In the absence of the server, user will get a choice to choose other profile or map provider