add fingerprinting for styles with the multi-language support
This resolves race condition when one style is requested immediately after
another while initializing the application. In this case, it is frequent when
style JSON of the first arrives after the second style has been requested
already. Without unique style ID, it would be assigned to the second
style and then used instead of the correct one. Such ID is required only
for styles that are processed by Pure Maps, such as Mapbox ones