reorganize cmake for better qt creator support
Source files are defined in the main folder making it possible to show
them all in the same hierarchy under Qt Creator.

Some reordering of cmake file content was done to push user-specified options to the top
and automatic processing to the later stages. Where feasible, install was done by file
listings as generated by cmake.